docker file image is edited

+++ backend/routes/view.py
@@ -7,6 +7,8 @@ db = Database()
 
 @views_bp.route('/drop', methods=['DELETE'])
 def drop():
+     if os.getenv('FLASK_ENV') != 'development':
+        return jsonify({'error': 'Forbidden'}), 403
     db.set_sql_file('db/queries/drop.sql')
     success, response = db.drop_table(db.get_sql_file())
     if(response == 200):
@@ -25,6 +27,8 @@ def drop():
     
 @views_bp.route('/')
 def welcome():
+    if os.getenv('FLASK_ENV') != 'development':
+        return jsonify({'error': 'Forbidden'}), 403
     db.set_sql_file('db/queries/initialize.sql')
     db.create_tables(db.get_sql_file())
     return jsonify({'message': 'Backend System is Established!'}), 200
